+++ src/rybak/_types.py
@@ -6,8 +6,3 @@ TemplateData: TypeAlias = Mapping[str, Any]
 
 Item = TypeVar('Item')
 LoopOverFn: TypeAlias = Callable[[Iterable[Item]], Item]
-RenderFn: TypeAlias = Callable[[str, str, TemplateData], None]
-
-
-class RenderError(Exception):
-    pass