+++ src/rybak/adapter.py
@@ -1,16 +1,11 @@
 import abc
 from pathlib import Path
-from typing import Any
 
 from ._types import LoopOverFn, TemplateData
 from .pycompat import Traversable
 
 
 class RendererAdapter(abc.ABC):
-    @abc.abstractmethod
-    def __init__(self, **kwargs: Any) -> None:
-        pass
-
     @abc.abstractmethod
     def render_str(self, template: str, data: TemplateData, loop_over: LoopOverFn) -> str:
         pass
@@ -23,3 +18,7 @@ class RendererAdapter(abc.ABC):
     @abc.abstractmethod
     def template_root(self) -> Traversable:
         pass
+
+
+class RenderError(Exception):
+    pass